index.html main page update
moved index.html from homepage folder to the main folder - clicking on main website link ([insert]) takes you to the homepage rather than README.md

deleted "play" folder, and moved relevant code (scroll wheel that Kam worked on, for instance, or the design of the main page) into the main index.html or to homepage.css, depending.

changed some references of the code on the main page to reference the correct place in the code given the above changes

this is the big one - I added the preloader screen. I needed to use chatGPT for some of it, and it currently looks ... pretty rough. but now when you open the website, you are met with an image of an adoption center, and when you click you are brought to our homepage.